backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-anthropic
langchain-google-genai
langchain-ollama
ddgs
duckduckgo-search
pydantic
httpx
python-dotenv
langchain-nvidia-ai-endpoints
langchain-openai
langgraph
pyautogen>=0.10.0
autogen-ext[openai]
pytest
slowapi
fredapi
